+++ models/staging/stg_obsidian_links.sql
@@ -1,0 +1,7 @@
+select
+  source_path,
+  block_id,
+  link_target,
+  link_text,
+  line_number
+from {{ source('obsidian', 'base_obsidian_links') }}